added notation interaction
