+++ mu4/domain/notation/inotationinteraction.h
@@ -16,8 +16,8 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 //=============================================================================
-#ifndef MU_DOMAIN_INOTATIONINPUTCONTROLLER_H
-#define MU_DOMAIN_INOTATIONINPUTCONTROLLER_H
+#ifndef MU_DOMAIN_INOTATIONINTERACTION_H
+#define MU_DOMAIN_INOTATIONINTERACTION_H
 
 #include <QPointF>
 #include <functional>
@@ -25,20 +25,38 @@
 #include "async/notification.h"
 
 #include "notationtypes.h"
+#include "inotationinputstate.h"
+#include "inotationselection.h"
 
 namespace mu {
 namespace domain {
 namespace notation {
-class INotationInputController
+class INotationInteraction
 {
 public:
-    virtual ~INotationInputController() = default;
+    virtual ~INotationInteraction() = default;
 
+    // Put notes
+    virtual void startNoteEntry() = 0;
+    virtual void endNoteEntry() = 0;
+    virtual void padNote(const Pad& pad) = 0;
+    virtual void putNote(const QPointF& pos, bool replace, bool insert) = 0;
+    virtual INotationInputState* inputState() const = 0;
+    virtual async::Notification inputStateChanged() const = 0;
+
+    // Shadow note
+    virtual void showShadowNote(const QPointF& p) = 0;
+    virtual void hideShadowNote() = 0;
+    virtual void paintShadowNote(QPainter* p) = 0;
+
+    // Select
     virtual Element* hitElement(const QPointF& pos, float width) const = 0;
+    virtual void select(Element* e, SelectType type, int staffIdx = 0) = 0;
+    virtual INotationSelection* selection() const = 0;
+    virtual async::Notification selectionChanged() const = 0;
 
-    // drag
+    // Drag
     using IsDraggable = std::function<bool (const Element*)>;
-
     virtual bool isDragStarted() const = 0;
     virtual void startDrag(const std::vector<Element*>& elems, const QPointF& eoffset, const IsDraggable& isDrag) = 0;
     virtual void drag(const QPointF& fromPos, const QPointF& toPos, DragMode mode) = 0;
@@ -49,4 +67,4 @@ public:
 }
 }
 
-#endif // MU_DOMAIN_INOTATIONINPUTCONTROLLER_H
+#endif // MU_DOMAIN_INOTATIONINTERACTION_H
