+++ mu4/scenes/notation/view/notationpaintview.cpp
@@ -84,11 +84,12 @@ void NotationPaintView::open()
     });
 
     onInputStateChanged();
-    m_notation->inputStateChanged().onNotify(this, [this]() {
+    INotationInteraction* ninteraction = notationInteraction();
+    ninteraction->inputStateChanged().onNotify(this, [this]() {
         onInputStateChanged();
     });
 
-    m_notation->selectionChanged().onNotify(this, [this]() {
+    ninteraction->selectionChanged().onNotify(this, [this]() {
         onSelectionChanged();
     });
 
@@ -108,7 +109,7 @@ void NotationPaintView::onViewSizeChanged()
 
 void NotationPaintView::onInputStateChanged()
 {
-    INotationInputState* is = m_notation->inputState();
+    INotationInputState* is = notationInteraction()->inputState();
     if (is->isNoteEnterMode()) {
         setAcceptHoverEvents(true);
     } else {
@@ -120,7 +121,7 @@ void NotationPaintView::onInputStateChanged()
 
 void NotationPaintView::onSelectionChanged()
 {
-    QRectF selRect = m_notation->selection()->canvasBoundingRect();
+    QRectF selRect = notationInteraction()->selection()->canvasBoundingRect();
     if (!selRect.isValid()) {
         return;
     }
@@ -135,12 +136,12 @@ bool NotationPaintView::isNoteEnterMode() const
         return false;
     }
 
-    return m_notation->inputState()->isNoteEnterMode();
+    return notationInteraction()->inputState()->isNoteEnterMode();
 }
 
 void NotationPaintView::showShadowNote(const QPointF& pos)
 {
-    m_notation->showShadowNote(pos);
+    notationInteraction()->showShadowNote(pos);
     update();
 }
 
@@ -153,7 +154,6 @@ void NotationPaintView::paint(QPainter* p)
 
     if (m_notation) {
         m_notation->paint(p, rect);
-        m_notation->paintShadowNote(p);
     } else {
         p->drawText(10, 10, "no notation");
     }
@@ -303,15 +303,23 @@ QPoint NotationPaintView::toPhysical(const QPoint& p) const
     return m_matrix.map(p);
 }
 
-std::shared_ptr<INotation> NotationPaintView::notation() const
-{
-    return m_notation;
-}
-
 bool NotationPaintView::isInited() const
 {
     if (m_notation) {
         return true;
     }
     return false;
+}
+
+std::shared_ptr<INotation> NotationPaintView::notation() const
+{
+    return m_notation;
+}
+
+mu::domain::notation::INotationInteraction* NotationPaintView::notationInteraction() const
+{
+    if (m_notation) {
+        return m_notation->interaction();
+    }
+    return nullptr;
 }
